Loop in a linked list
Given the starting node (head) of a singly linked list, your task is to find whether the linked list contains a loop. A loop in a linked list exists when a node can be revisited by continuously following the next pointers.

You should return true if the linked list forms a loop (sometimes referred to as a circular list or a cyclic list) and false if it does not.